feat(PKH): rename function to match the file name and change embed to attach for better readability
